Rename and move translations file
