+++ src_ts/utils/translate.ts
@@ -1,4 +1,4 @@
-import translations from '../assets/translations/all';
+import translations from '../assets/translations';
 
 export function getTranslation(lang: string, key: string) {
   try {
